修改下单接口

+++ mini-api/src/main/java/com/gongsi/mini/enums/ActivityStatusEn.java
@@ -4,20 +4,20 @@ package com.gongsi.mini.enums;
  * Created by 吴宇 on 2018-05-24.
  */
 public enum  ActivityStatusEn {
-    ACTIVE(1,"进行中"),
-    END(5,"结束"),
+    ACTIVE((byte)1,"进行中"),
+    END((byte)5,"结束"),
     ;
 
-    ActivityStatusEn(Integer code, String desc) {
+    ActivityStatusEn(Byte code, String desc) {
         this.code = code;
         this.desc = desc;
     }
 
-    private Integer code;
+    private Byte code;
 
     private String desc;
 
-    public Integer getCode() {
+    public Byte getCode() {
         return code;
     }
